+++ templates/mca/multi/MCA.Api/Program.cs
@@ -2,11 +2,9 @@
 using Serilog;
 #endif
 using Scalar.AspNetCore;
-using MCA.Application;
 using MCA.Domain.Interfaces;
 using MCA.Infrastructure.Data;
 using MCA.Infrastructure.Repositories;
-using MCA.Application.Commands;
 using MCA.Application.Handlers;
 using MCA.Api.Endpoints;
 using Microsoft.EntityFrameworkCore;
@@ -30,6 +28,7 @@ using MinimalCleanArch.Security.Extensions;
 using MinimalCleanArch.Extensions.Caching;
 #endif
 #if (UseMessaging)
+using MCA.Application;
 using MinimalCleanArch.Messaging.Extensions;
 #endif
 #if (UseAudit)
@@ -211,8 +210,8 @@ else
 #endif
 
 #if (UseStorage)
-// Blob storage (Azure Blob Storage / Azurite). Configure BlobStorage:* in appsettings or env.
-builder.Services.AddAzureBlobStorage(builder.Configuration);
+// Blob storage: BlobStorage:Provider = Azure (default, Azurite) or R2 (Cloudflare).
+builder.Services.AddBlobStorage(builder.Configuration);
 #endif
 
 #if (UseSecurity)
@@ -358,9 +357,9 @@ var app = builder.Build();
 if (app.Environment.IsDevelopment())
 {
     app.MapOpenApi();
+#if (UseAuth)
     app.MapScalarApiReference(options =>
     {
-#if (UseAuth)
         var webClientSecret = app.Configuration["OpenIddict:Clients:Web:Secret"];
         var webClientId = app.Configuration["OpenIddict:Clients:Web:ClientId"] ?? "mca-web-client";
         options.AddPasswordFlow("oauth2", flow =>
@@ -378,9 +377,11 @@ if (app.Environment.IsDevelopment())
             };
         });
         options.AddPreferredSecuritySchemes(new[] { "oauth2" });
-        options.WithPersistentAuthentication();
-#endif
+        options.EnablePersistentAuthentication();
     });
+#else
+    app.MapScalarApiReference();
+#endif
 }
 
 #if (UseMcaApiBootstrap)